functions/firebase/python/requirements.txt
flask
yahooquery
google-cloud-logging